Update button positioning on setConsent, rename extensionViewForm
