+++ src/view/components/partialForm.jsx
@@ -1,3 +1,4 @@
+import PropTypes from "prop-types";
 import usePartialForm from "../utils/usePartialForm";
 
 const PartialForm = ({
@@ -22,6 +23,15 @@ const PartialForm = ({
     return null;
   }
   return render({ initInfo });
+};
+
+PartialForm.propTypes = {
+  getInitialValues: PropTypes.func.isRequired,
+  getSettings: PropTypes.func.isRequired,
+  validateFormikState: PropTypes.func,
+  formikStateValidationSchema: PropTypes.object,
+  validateNonFormikState: PropTypes.func,
+  name: PropTypes.string.isRequired
 };
 
 // TODO add prop types validation
